requirements.txt
playwright==1.51.0
pytest
pytest-playwright
pytest-xdist
allure-pytest